fix:Modified variable type
Changed score's variable type.

+++ database/migrations/2021_10_23_0920012_create_objectives_table.php
@@ -16,7 +16,7 @@ class CreateObjectivesTable extends Migration
         Schema::create('objectives', function (Blueprint $table) {
             $table->increments('id');
             $table->text('result')->comment('成果詳細');
-            $table->integer('score')
+            $table->float('score')
                 ->nullable()
                 ->comment('個別スコア');
             $table->integer('okr_id')
